Further refactoring of package structure
